Allow MIDI port to be specified by number
Command-line handling on Windows is terrible and broken. It's not
reasonable to pass a string with spaces (such as a MIDI port name like
'2- FF 400 MIDI 2 2' )and expect it to be received properly.

Instead, enumerate the MIDI ports and allow them to be selected by
number with the argument --midi-port-index.

If the MIDI port cannot be found or none is given, the error shows a
list of all the MIDI ports. This is helpful.